Refactored Audio Reording/Playback Into Services.
